requirements.txt
groq
fastapi
uvicorn
python-dotenv
PyPDF2
python-multipart
requests
python-docx